adds proper test (even if basic) and fixes undefined issue
